Work on a better solution for uva problem 11947
This has to do with getting the signs depending on the
date and month.

Instead of using a bunch of 'if' statments, I want to use
some smart logic to get the answer.
So far it's working with the input provided from the question,
but it fails the input from uDebug.

Need to rethink it.